+++ ec2_ssh.py
@@ -85,7 +85,12 @@ def host():
 def get_dns_names(tag, value):
     conn = boto3.client('ec2')
 
-    filters = []
+    filters = [
+        {
+            'Name': 'instance-state-name',
+            'Values': ['running'],
+        }
+    ]
     if value:
         filters.append({
             'Name': 'tag:' + tag,